+++ platform/window.h
@@ -29,10 +29,7 @@ struct p_window_meta {
 struct p_window * p_window_open(const unsigned char *title,
     const rect_t *area, const u32 flags);
 
-void p_window_bind_fb(struct p_window *win, struct pixel_flat_data *fb);
-void p_window_unbind_fb(struct p_window *win);
-
-void p_window_render(struct p_window *win);
+void p_window_render(struct p_window *win, struct pixel_flat_data *fb);
 
 void p_window_close(struct p_window **win_p);