+++ platform/windows/window-internal.h
@@ -27,10 +27,6 @@ struct p_window {
     u32 screen_w, screen_h;
 
     HDC dc; /* The device context of the window */
-
-    /* The framebuffer that the `p_window_render` displays */
-    struct pixel_flat_data *bound_fb;
-    BITMAPINFO bound_fb_bmi; /* The bitmap info of the bound framebuffer */
 };
 
 struct window_init {