+++ render/rctx.c
@@ -15,6 +15,8 @@
 
 #define MODULE_NAME "rctx"
 
+static void swap_buffers(struct r_ctx *rctx);
+
 struct r_ctx * r_ctx_init(struct p_window *win, enum r_type type, u32 flags)
 {
     u_check_params(win != NULL);
@@ -37,17 +39,20 @@ struct r_ctx * r_ctx_init(struct p_window *win, enum r_type type, u32 flags)
     if (p_window_get_meta(ctx->win, &ctx->win_meta))
         goto_error("Failed to get window metadata!");
 
-    s_log_debug("win_meta->w: %u, win_meta->h: %u", ctx->win_meta.w, ctx->win_meta.h);
+    s_log_debug("win_meta->w: %u, win_meta->h: %u",
+        ctx->win_meta.w, ctx->win_meta.h);
 
-    ctx->pixels.buf = calloc(ctx->win_meta.w * ctx->win_meta.h, sizeof(pixel_t));
-    s_assert(ctx->pixels.buf != NULL, "calloc() failed for the pixel buffer");
+    ctx->buffers[0].buf = calloc(ctx->win_meta.w * ctx->win_meta.h, sizeof(pixel_t));
+    ctx->buffers[1].buf = calloc(ctx->win_meta.w * ctx->win_meta.h, sizeof(pixel_t));
+    s_assert(ctx->buffers[0].buf != NULL && ctx->buffers[1].buf != NULL,
+        "calloc() failed for the pixel buffer(s)");
 
-    ctx->pixels.w = ctx->win_meta.w;
-    ctx->pixels.h = ctx->win_meta.h;
-    u_rect_from_pixel_data(&ctx->pixels, &ctx->pixels_rect);
+    ctx->buffers[0].w = ctx->buffers[1].w = ctx->win_meta.w;
+    ctx->buffers[0].h = ctx->buffers[1].h = ctx->win_meta.h;
+    u_rect_from_pixel_data(&ctx->buffers[0], &ctx->pixels_rect);
 
-    p_window_bind_fb(win, &ctx->pixels);
-
+    ctx->render_buffer = &ctx->buffers[0];
+    ctx->blit_buffer = &ctx->buffers[1];
     ctx->current_color = BLACK_PIXEL;
 
     /* Render 1 empty frame on init
@@ -67,8 +72,8 @@ void r_ctx_destroy(struct r_ctx **ctx_p)
     if (ctx_p == NULL || *ctx_p == NULL) return;
     struct r_ctx *ctx = *ctx_p;
 
-    p_window_unbind_fb(ctx->win);
-    free(ctx->pixels.buf);
+    free(ctx->buffers[0].buf);
+    free(ctx->buffers[1].buf);
     u_nzfree(ctx_p);
 }
 
@@ -85,11 +90,25 @@ void r_ctx_set_color(struct r_ctx *ctx, color_RGBA32_t color)
 void r_flush(struct r_ctx *ctx)
 {
     u_check_params(ctx != NULL);
-    p_window_render(ctx->win);
+    swap_buffers(ctx);
+    p_window_render(ctx->win, ctx->blit_buffer);
 }
 
 void r_reset(struct r_ctx *ctx)
 {
     u_check_params(ctx != NULL);
-    memset(ctx->pixels.buf, 0, ctx->pixels.w * ctx->pixels.h * sizeof(pixel_t));
+    memset(ctx->render_buffer->buf, 0,
+        ctx->render_buffer->w * ctx->render_buffer->h * sizeof(pixel_t)
+    );
+}
+
+static void swap_buffers(struct r_ctx *rctx)
+{
+    if (rctx->render_buffer == &rctx->buffers[0]) {
+        rctx->blit_buffer = &rctx->buffers[0];
+        rctx->render_buffer = &rctx->buffers[1];
+    } else {
+        rctx->blit_buffer = &rctx->buffers[1];
+        rctx->render_buffer = &rctx->buffers[0];
+    }
 }